Brought library up to node 0.8 standards, added package.json for inclusion in npm if author so wishes
