+++ Makefile
@@ -1,0 +1,6 @@
+TESTS = $(shell find test -name "*test.js")
+
+test:
+	node $(TESTS)
+
+.PHONY: test
